pyproject.toml
[tool.poetry]
name = "weekend-activity"
version = "0.1.0"
description = "Track and summarize weekend GitHub activity"
authors = ["Gajesh Naik"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/Gajesh2007/weekend-activity"
repository = "https://github.com/Gajesh2007/weekend-activity"
documentation = "https://github.com/Gajesh2007/weekend-activity#readme"
keywords = ["github", "activity", "tracking", "weekend", "slack"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Version Control :: Git",
]
packages = [
    { include = "weekend_activity", from = "src" }
]

[tool.poetry.dependencies]
python = "^3.9"
PyGithub = "^2.1.1"
python-dateutil = "^2.8.2"
pytz = "^2024.1"
pyyaml = "^6.0.1"
requests = "^2.31.0"
click = "^8.1.7"
rich = "^13.7.0"
SQLAlchemy = "^2.0.25"
alembic = "^1.13.1"
openai = "^1.10.0"
python-dotenv = "^1.0.1"
aiosqlite = "^0.19.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
black = "^24.1.1"
flake8 = "^7.0.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
mypy = "^1.8.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
weekend-activity = "weekend_activity.cli:main"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "--cov=weekend_activity --cov-report=term-missing"
